backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
pydantic-settings>=2.6.0
trafilatura>=1.9.0
yt-dlp>=2025.1.0
